Initial transfer to organization repository
Add all work so far and consolidate into a single repo namespace.
This repository consists of `nine`, which is a set of routines for
working with lower-level 9P / byte streams, and `ramfs`, which is
a complete implementation of 9P for an in-memory string store.
Ramfs isn't completely ready for release yet, because it's buggy
and hasn't been fully tested.

Also add a simple client, which runs in a separate thread and
connects to a running server over loopback. The current iteration
is just a pre-programmed set of operations that are known to work,
but this is a future line of work we've discussed with the team.

Signed-off-by: Jay Lang <jaytlang@mit.edu>