Nits: remove 9P version from configuration and establish as constant
Use of the NineVersion string was inconsistent - sometimes it was
passed in a struct and other times the global constant was used.
Fix this, so that the global constant is used and we assume no other
9P versions are supported.

In addition, enable version checking, so that if the server does
not receive the right version string (i.e. the global constant),
reply with "Unknown" per the 9P spec. Also don't register the ID,
such that the session is not formally established.

Signed-off-by: Jay Lang <jaytlang@mit.edu>