Fix a bug which caused directory creation to fail
During validation of permission bits, we used to AND the permission
bits with 0777/0666 in order to set a default behavior. However, we
never reverted back to the original 32 bit mode sent by the client,
which resulted in the masking out of flags that dictated file
behavior. Fix this by altogether getting rid of default behavior:
if you can create a file, allow creation with any flags you please
as long as it isn't illegal.

Signed-off-by: Jay Lang <jaytlang@mit.edu>