Ensure that the root kchan is unique and fix clunking
Ensure that all fids start with 1, so the root kchan holds onto
zero. Since the root kchan is a dummy, now kchancmp can work on
it to ensure it isn't accidentally used in an fcall. Some caller
responsibility is obviously required in order to not accidentally
clunk something, but I think this strikes a good compromise.

Besides the special cwd/root cases, add a goto statement in the path
eval function to ensure that any _other_ fid it returns is clunkable.

Signed-off-by: Jay Lang <jaytlang@mit.edu>