Add close draft, prepare to fix names
Turns out the dot dot problem was solved only because we were
mutating the mount table. We need the mount table to stay static
so that close and unmount work down the road...accordingly, fix
up and prepare to refactor kchans so that they store physical location
as well as cnames.

Signed-off-by: Jay Lang <jaytlang@mit.edu>